+++ src/compute-engine/latex-syntax/dictionary/definitions-arithmetic.ts
@@ -2009,7 +2009,7 @@ export const DEFINITIONS_ARITHMETIC: LatexDictionary = [
     precedence: COMPARISON_PRECEDENCE,
     parse: (parser) => {
       const rhs = parser.parseGroup() ?? parser.parseToken();
-      return ['Mod', missingIfEmpty(rhs)];
+      return ['Mod', parser.error('missing', parser.index), missingIfEmpty(rhs)];
     },
   },
   {